# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/wallet/topup/initiate": {
      "post": {
        "operationId": "WalletController_initiateTopup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateTopupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
...
    "/api/v1/wallet/topup/manual": {
      "post": {
        "operationId": "WalletController_manualTopup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateTopupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
...
  },
  "info": {
    "title": "SabiNote API",
    "description": "SabiNote Education Platform API — Grounded Nigerian Lesson & Curriculum Engine",
    "version": "2.0.0-baseline",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "InitiateTopupDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}